# Improve Clamp() performance