--- a/src/Core/src/Graphics/CompareExtensions.cs
+++ b/src/Core/src/Graphics/CompareExtensions.cs
@@ -1,9 +1,11 @@
-﻿using System;
+using System;
+using System.Runtime.CompilerServices;
 
 namespace Microsoft.Maui.Graphics
 {
 	internal static class CompareExtensions
 	{
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static T Clamp<T>(this T value, T min, T max) where T : IComparable<T>
 		{
 			if (max.CompareTo(min) < 0)