--- a/src/Core/src/Graphics/CompareExtensions.cs
+++ b/src/Core/src/Graphics/CompareExtensions.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Runtime.CompilerServices;
 
 namespace Microsoft.Maui.Graphics
@@ -6,7 +6,7 @@ namespace Microsoft.Maui.Graphics
 	internal static class CompareExtensions
 	{
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static T Clamp<T>(this T value, T min, T max) where T : IComparable<T>
+		public static DateTime Clamp(this DateTime value, DateTime min, DateTime max)
 		{
 			if (max.CompareTo(min) < 0)
 				return max;