--- a/src/Graphics/src/Graphics/NumericExtensions.cs
+++ b/src/Graphics/src/Graphics/NumericExtensions.cs
@@ -1,4 +1,6 @@
+#if NETSTANDARD2_1_OR_GREATER
 using System;
+#endif
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
 
@@ -10,6 +12,9 @@ namespace Microsoft.Maui.Graphics
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static float Clamp(this float self, float min, float max)
 		{
+#if NETCOREAPP3_0_OR_GREATER
+			return float.Clamp(self, min, max);
+#else
 			if (max < min)
 			{
 				return max;
@@ -24,11 +29,15 @@ namespace Microsoft.Maui.Graphics
 			}
 
 			return self;
+#endif
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static double Clamp(this double self, double min, double max)
 		{
+#if NETCOREAPP3_0_OR_GREATER
+			return double.Clamp(self, min, max);
+#else
 			if (max < min)
 			{
 				return max;
@@ -43,11 +52,15 @@ namespace Microsoft.Maui.Graphics
 			}
 
 			return self;
+#endif
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static int Clamp(this int self, int min, int max)
 		{
+#if NETSTANDARD2_1_OR_GREATER
+			return Math.Clamp(self, min, max);
+#else
 			if (max < min)
 			{
 				return max;
@@ -62,6 +75,7 @@ namespace Microsoft.Maui.Graphics
 			}
 
 			return self;
+#endif
 		}
 	}
 }