… src/Graphics/src/Graphics/Properties/AssemblyInfo.cs