The current Clamp() extension based on IComparable is about twice as slow for floats and doubles, and was being used in at least 35 places.

It is just an internal function that was not used on anything interesting that implements IComparable. It only was used on int, double, and float types. And one DateTime.

The most critical path for clamp is probably [Color.cs](https://github.com/dotnet/maui/blob/b4bc1c85e5b2deb19a75cf35ab8b12e916f1c1e9/src/Graphics/src/Graphics/Color.cs#L4) 

There already was another NumericExtensions class in the Graphics assembly for dealing with those types. Replicating the use of AggressiveInlining on the IComparable implementation, to align with NumericExtensions, doesn't change much about the results.

There is also [Math.Clamp()](https://learn.microsoft.com/en-us/dotnet/api/system.math.clamp), but it only exists on .NET Standard 2.1 and above. It allocates less for ints, but NumericExtensions is slightly better otherwise.

There is no MathF.Clamp() for floats, unfortunately.

| Method                         | Mean      | Error    | StdDev   | Median    | Allocated |
|------------------------------- |----------:|---------:|---------:|----------:|----------:|
| ComparableClampInt    |  28.66 ms | 0.503 ms | 0.516 ms |  28.64 ms |      57 B |
| NativeClampInt        |  32.23 ms | 1.564 ms | 4.612 ms |  29.56 ms |      63 B |
| MathClampInt          |  27.72 ms | 0.253 ms | 0.211 ms | 27.75 ms  |      12 B |

| Method                         | Mean      | Error    | StdDev   | Median    | Allocated |
|------------------------------- |----------:|---------:|---------:|----------:|----------:|
| ComparableClampFloat  | 100.16 ms | 1.978 ms | 2.117 ms | 100.74 ms |     100 B |
| NativeClampFloat      |  49.73 ms | 0.699 ms | 0.654 ms |  49.72 ms |      36 B |
| MathClampFloat        |  50.00 ms | 0.773 ms | 0.685 ms | 49.95 ms  |      40 B |

| Method                         | Mean      | Error    | StdDev   | Median    | Allocated |
|------------------------------- |----------:|---------:|---------:|----------:|----------:|
| ComparableClampDouble |  97.51 ms | 1.717 ms | 1.341 ms |  97.53 ms |      80 B |
| NativeClampDouble     |  47.52 ms | 0.498 ms | 0.466 ms |  47.60 ms |      40 B |
| MathClampDouble       |  50.25 ms | 0.817 ms | 0.764 ms | 50.35 ms  |      40 B |