@symbiogenesis 2024-01-22
Aligning the ambiguous namespace resolution issue with PR https://github.com/dotnet/maui/pull/19965, which should be cleaner